+++ src/database/employeesAccess.js
@@ -1,0 +1,18 @@
+import database from '../config/firebaseConfig';
+
+export const getEmployeesByPermission = () => {
+    return database.collection('employees').where('admin', '==', false).get()
+        .then(querySnapshot => {
+            let employeeArray = []
+            querySnapshot.forEach(employee => {
+                employeeArray.push({ ...employee.data(), dbId: employee.id })
+            });
+            return Promise.resolve({ employees: employeeArray});
+        })
+        .catch(err => console.log("Error getting documents: ", err));
+};
+
+export const setAdminStatus = (employeeId, bool) => {
+    return database.collection('employees').doc(employeeId)
+        .update({ admin: bool })
+};
